+++ lib/src/init_dialog.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
 import 'package:flutter_smart_dialog/src/helper/navigator_observer.dart';
 import 'package:flutter_smart_dialog/src/util/view_utils.dart';
 import 'package:flutter_smart_dialog/src/widget/toast_widget.dart';
@@ -19,6 +20,7 @@ class FlutterSmartDialog extends StatefulWidget {
     this.toastBuilder,
     this.loadingBuilder,
     this.styleBuilder,
+    this.initType,
   }) : super(key: key);
 
   final Widget? child;
@@ -31,6 +33,9 @@ class FlutterSmartDialog extends StatefulWidget {
 
   ///Compatible with cupertino style
   final FlutterSmartStyleBuilder? styleBuilder;
+
+  ///inti type
+  final Set<SmartInitType>? initType;
 
   @override
   _FlutterSmartDialogState createState() => _FlutterSmartDialogState();
@@ -50,6 +55,8 @@ class FlutterSmartDialog extends StatefulWidget {
     FlutterSmartLoadingBuilder? loadingBuilder,
     //Compatible with cupertino style
     FlutterSmartStyleBuilder? styleBuilder,
+    //init type
+    Set<SmartInitType>? initType,
   }) {
     MonitorPopRoute.instance;
 
@@ -59,6 +66,7 @@ class FlutterSmartDialog extends StatefulWidget {
               toastBuilder: toastBuilder,
               loadingBuilder: loadingBuilder,
               styleBuilder: styleBuilder,
+              initType: initType,
               child: child,
             )
           : builder(
@@ -67,6 +75,7 @@ class FlutterSmartDialog extends StatefulWidget {
                 toastBuilder: toastBuilder,
                 loadingBuilder: loadingBuilder,
                 styleBuilder: styleBuilder,
+                initType: initType,
                 child: child,
               ),
             );
@@ -75,6 +84,9 @@ class FlutterSmartDialog extends StatefulWidget {
 }
 
 class _FlutterSmartDialogState extends State<FlutterSmartDialog> {
+  late FlutterSmartStyleBuilder styleBuilder;
+  late Set<SmartInitType> initType;
+
   @override
   void initState() {
     ViewUtils.addSafeUse(() {
@@ -99,34 +111,52 @@ class _FlutterSmartDialogState extends State<FlutterSmartDialog> {
     proxy.toastBuilder = widget.toastBuilder ?? defaultToast;
     proxy.loadingBuilder = widget.loadingBuilder ?? defaultLoading;
 
+    // init param
+    styleBuilder = widget.styleBuilder ??
+        (Widget child) {
+          return Material(color: Colors.transparent, child: child);
+        };
+    initType = widget.initType ??
+        {SmartInitType.custom, SmartInitType.attach, SmartInitType.loading, SmartInitType.toast};
+
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return widget.styleBuilder == null
-        ? Material(color: Colors.transparent, child: _buildOverlay())
-        : widget.styleBuilder!.call(_buildOverlay());
-  }
-
-  Widget _buildOverlay() {
-    return Overlay(initialEntries: [
+    return styleBuilder(Overlay(initialEntries: [
       //main layout
       OverlayEntry(
         builder: (BuildContext context) => widget.child ?? Container(),
       ),
 
       //provided separately for custom dialog
-      OverlayEntry(builder: (BuildContext context) {
-        DialogProxy.contextOverlay = context;
-        return Container();
-      }),
+      if (initType.contains(SmartInitType.custom) && !initType.contains(SmartInitType.attach))
+        OverlayEntry(builder: (BuildContext context) {
+          DialogProxy.contextCustom = context;
+          return Container();
+        }),
+
+      //provided separately for attach dialog
+      if (initType.contains(SmartInitType.attach) && !initType.contains(SmartInitType.custom))
+        OverlayEntry(builder: (BuildContext context) {
+          DialogProxy.contextAttach = context;
+          return Container();
+        }),
+
+      //provided separately for custom dialog and attach dialog
+      if (initType.contains(SmartInitType.custom) && initType.contains(SmartInitType.attach))
+        OverlayEntry(builder: (BuildContext context) {
+          DialogProxy.contextCustom = context;
+          DialogProxy.contextAttach = context;
+          return Container();
+        }),
 
       //provided separately for loading
-      DialogProxy.instance.entryLoading,
+      if (initType.contains(SmartInitType.loading)) DialogProxy.instance.entryLoading,
 
       //provided separately for toast
-      DialogProxy.instance.entryToast,
-    ]);
+      if (initType.contains(SmartInitType.toast)) DialogProxy.instance.entryToast,
+    ]));
   }
 }