+++ lib/src/custom/custom_dialog.dart
@@ -223,11 +223,7 @@ class CustomDialog extends BaseDialog {
       mainDialog = singleDialogInfo.dialog.mainDialog;
       dialogInfo = singleDialogInfo;
     } else {
-      if (displayTime != null) {
-        tag = tag ?? '${displayTime.hashCode + Random().nextInt(666666) + Random().nextDouble()}';
-      } else if (bindWidget != null) {
-        tag = tag ?? "${bindPage.hashCode}";
-      }
+      tag = tag ?? '${hashCode + Random().nextDouble()}';
 
       // handle dialog stack
       dialogInfo = DialogInfo(
@@ -290,8 +286,8 @@ class CustomDialog extends BaseDialog {
     var debounceTime = type == DialogType.dialog
         ? SmartDialog.config.custom.debounceTime
         : SmartDialog.config.attach.debounceTime;
-    var prohibit =
-        proxy.dialogLastTime != null && now.difference(proxy.dialogLastTime!) < debounceTime;
+    var prohibit = proxy.dialogLastTime != null &&
+        now.difference(proxy.dialogLastTime!) < debounceTime;
     proxy.dialogLastTime = now;
     if (prohibit) return false;
 
@@ -317,7 +313,9 @@ class CustomDialog extends BaseDialog {
     bool force = false,
     CloseType closeType = CloseType.normal,
   }) {
-    if (type == DialogType.dialog || type == DialogType.custom || type == DialogType.attach) {
+    if (type == DialogType.dialog ||
+        type == DialogType.custom ||
+        type == DialogType.attach) {
       return _closeSingle<T>(
         type: type,
         tag: tag,